services/sentiment/requirements.txt
google-generativeai==0.7.2
psycopg2-binary==2.9.9